Add params to the select form partial to allow manual calling.
Most other common form partials can be called manually:

  {{> core/forms/common/email label='Email Address' name='email' required=true}}

But the select partial can not have its options generated unless your data is
available in a very specific format.

This adds a new `select_type` parameter that can be used to call the partial
with different types of data.

`range` generates options based on `from` and `to` parameters.

`array` generates options using a single array of values, setting the value
and label to the same thing.

And if you don't define the `select_type` parameter, it works as before,
generating options using an array of label-value pairs.